requirements.txt
google-genai
opencv-python
numpy
pillow
